+++ src/interfaces/SwadeGame.ts
@@ -3,6 +3,8 @@ import RollDialog from '../module/apps/RollDialog';
 import SettingConfigurator from '../module/apps/SettingConfigurator';
 import SwadeDocumentTweaks from '../module/apps/SwadeDocumentTweaks';
 import CharacterSummarizer from '../module/CharacterSummarizer';
+import Benny from '../module/dice/Benny';
+import WildDie from '../module/dice/WildDie';
 import ItemChatCardHelper from '../module/ItemChatCardHelper';
 import * as migrations from '../module/migration';
 import CharacterSheet from '../module/sheets/official/CharacterSheet';
@@ -23,6 +25,10 @@ export default interface SwadeGame {
     SwadeDocumentTweaks: typeof SwadeDocumentTweaks;
     AdvanceEditor: typeof AdvanceEditor;
     SettingConfigurator: typeof SettingConfigurator;
+  };
+  dice: {
+    Benny: typeof Benny;
+    WildDie: typeof WildDie;
   };
   CharacterSummarizer: typeof CharacterSummarizer;
   RollDialog: typeof RollDialog;